requirements.txt
flask 
beautifulsoup4 
requests 
sentence-transformers 
faiss-cpu 
langchain